State - Action Fusion
States and actions now exist together in 1 file.

In the new system there are only states which derive from ScriptableObject. That makes states customizable and reproducable. Which the old system also did but now everything about a state is in one file and inheritance architecture is now healthier and easier to implement and edit.

This commit is not ready. The game will not work in this. This is a checkpoint before the implementation of revised input system.